Fix rv policy for overloaded functions with custom policies.
Before, there was a problematic edge case: a function with a custom rv
policy set in the config for the unspecific function name, which has overloads
where one returns some reference type, and hence needs an rv policy,
but another overload that returns a non-reference type, would receive an
rv policy in the bindings for both overloads.

Now, the overload that does not require an rv policy to be set does not
get one in the bindings.